organizing code and fixed a bug
